capture, test: Move stubs to it's own file
